Fix and added Company and Hr
