+++ src/lekcja6_zadanie1_tablica_pracownikow1/Accountancy.java
@@ -1,13 +1,9 @@
 package lekcja6_zadanie1_tablica_pracownikow1;
 
 public class Accountancy {
-    public static void main(String[] args) {
 
-        Employee employee = new Employee();
-        Employee[] employees = new Employee[2];
-        employees[0] = employee.createEmployee();
-        employees[1] = employee.createEmployee();
-        double paymentsSum = employees[0].getPayment() + employees[1].getPayment();
-        System.out.println("Firma musi wydać na wypłaty dla pracowników: " + paymentsSum + "zł");
+    public void paymentsSum(Employee[] employees) {
+        double sum = employees[0].getPayment() + employees[1].getPayment();
+        System.out.println("Firma musi wydać na wypłaty dla pracowników: " + sum + "zł");
     }
 }
